// typescript-projects/utility-backups/src20250516BeforeErrorHandlingChange/interface/types.ts
export type TEbillsElectricityPaymentData = {
  provider: ProviderId;
  merchant: string;
  customerId: string;
  amount: number; // `amount` in tx doc root
};